# profesores-backend/openapi.yaml
openapi: 3.0.0

info:
  title: CRUD de Profesores UNINPAHU
  description: |
    Microservicio para administrar información de profesores de UNINPAHU.
    La API permite realizar las operaciones CRUD: crear, consultar,
    actualizar y eliminar profesores.
  version: 2.0.0

servers:
  - url: https://profesores-backend.onrender.com

paths:

  /api/profesores:

    get:
      summary: Consultar profesores
      description: |
        Permite obtener todos los profesores registrados, consultar un
        profesor mediante su ID o buscar profesores mediante su nombre.
        La búsqueda por nombre ignora mayúsculas, minúsculas y tildes.

      parameters:

        - name: id
          in: query
          required: false
          description: ID numérico del profesor que se desea consultar.
          schema:
            type: integer
            example: 1

        - name: nombre
          in: query
          required: false
          description: Nombre o parte del nombre. Debe tener al menos 3 caracteres.
          schema:
            type: string
            minLength: 3
            example: Andres

      responses:

        '200':
          description: Consulta realizada correctamente.
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    description: Lista de profesores.
                    items:
                      $ref: '#/components/schemas/Profesor'
                  - $ref: '#/components/schemas/Profesor'

        '400':
          description: Los parámetros enviados no tienen un formato válido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                id:
                  value:
                    error: El parámetro id debe ser un número
                nombre:
                  value:
                    error: El nombre debe tener al menos 3 caracteres

        '404':
          description: No se encontró el profesor solicitado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: No se encontraron profesores

        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Error interno del servidor


    post:
      summary: Crear un profesor
      description: |
        Crea un nuevo profesor en la colección de MongoDB.
        El ID es generado automáticamente por el microservicio.

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfesorInput'
            example:
              nombre: Elfar Didier Morantes Sánchez
              foto: https://example.com/elfar.jpg
              formacion: Especialista en Ingeniería de Software; Magíster en E-Learning
              profesion: Ingeniero de Software
              cargo: Docente
              area: Ingeniería de Software, desarrollo web y tecnologías de la información

      responses:

        '201':
          description: Profesor creado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profesor'

        '400':
          description: Faltan datos obligatorios o el cuerpo no contiene JSON válido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: El campo nombre es obligatorio

        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Error interno del servidor


    put:
      summary: Actualizar un profesor
      description: |
        Actualiza los datos de un profesor existente.
        Se debe enviar el ID y todos los campos del profesor.

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfesorUpdate'
            example:
              id: 6
              nombre: Elfar Didier Morantes Sánchez
              foto: https://example.com/elfar-actualizado.jpg
              formacion: Magíster en Ingeniería de Software
              profesion: Ingeniero de Software
              cargo: Docente
              area: Ingeniería de Software y desarrollo web

      responses:

        '200':
          description: Profesor actualizado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profesor'

        '400':
          description: El ID o alguno de los campos obligatorios no es válido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                id:
                  value:
                    error: El campo id debe ser un número
                campo:
                  value:
                    error: El campo nombre es obligatorio

        '404':
          description: No existe un profesor con el ID indicado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Profesor no encontrado

        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Error interno del servidor


    delete:
      summary: Eliminar un profesor
      description: Elimina de MongoDB el profesor cuyo ID sea indicado.

      parameters:

        - name: id
          in: query
          required: true
          description: ID numérico del profesor que se desea eliminar.
          schema:
            type: integer
            example: 6

      responses:

        '200':
          description: Profesor eliminado correctamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: Profesor eliminado correctamente

        '400':
          description: El parámetro id es obligatorio y debe ser numérico.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                faltante:
                  value:
                    error: El parámetro id es obligatorio
                invalido:
                  value:
                    error: El parámetro id debe ser un número

        '404':
          description: No existe un profesor con el ID indicado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Profesor no encontrado

        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Error interno del servidor


components:

  schemas:

    Profesor:
      type: object
      properties:
        id:
          type: integer
          example: 1

        nombre:
          type: string
          example: Omar Andrés Bonilla Acosta

        foto:
          type: string
          format: uri
          example: https://uninpahu.edu.co/imagen-profesor.jpg

        formacion:
          type: string
          example: Ingeniero en Sistemas

        profesion:
          type: string
          example: Ingeniero en Sistemas

        cargo:
          type: string
          example: Docente

        area:
          type: string
          example: Sistemas y Seguridad Informática


    ProfesorInput:
      type: object

      required:
        - nombre
        - foto
        - formacion
        - profesion
        - cargo
        - area

      properties:

        nombre:
          type: string
          example: Elfar Didier Morantes Sánchez

        foto:
          type: string
          format: uri
          example: https://example.com/elfar.jpg

        formacion:
          type: string
          example: Especialista en Ingeniería de Software; Magíster en E-Learning

        profesion:
          type: string
          example: Ingeniero de Software

        cargo:
          type: string
          example: Docente

        area:
          type: string
          example: Ingeniería de Software y desarrollo web


    ProfesorUpdate:
      allOf:

        - type: object

          required:
            - id

          properties:

            id:
              type: integer
              example: 6

        - $ref: '#/components/schemas/ProfesorInput'


    Error:
      type: object

      properties:

        error:
          type: string
          example: Error